refactor(mcp): extract tool test modules
